// src/types/generated/TextFile.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type TextFile = { path: string, content: string, 
/**
 * Detected encoding ("utf-8" or "latin-1").
 */
encoding: string, size: number, modified: number | null, permissions: number | null, 
/**
 * Line ending style detected ("lf" | "crlf").
 */
eol: string, };